Updated all dt's to be of type double. Got aspects implemented (including physics derivative)

+++ include/Aspect.h
@@ -10,7 +10,7 @@ public:
     Aspect(Entity *ent);
     virtual ~Aspect();
 
-    virtual void Tick(uint32_t dt);
+    virtual void Tick(double dt);
 
     Entity *entity;
 };
