+++ include/EntityMgr.h
@@ -14,7 +14,7 @@ public:
     ~EntityMgr();
 
     void Init();
-    void Tick(uint32_t dt);
+    void Tick(double dt);
     void Stop();
 
     Engine *engine;
